simplified and responsive SVG example: blocks.svg
Mostly generated through a convo with GPT-4o. It needed a lot of extra coaching
and walking through the steps of making cubes that are not transparent, then
making one on top of the other, etc.

By "responsive" I mean that the colors change depending on the selected website
theme.

In order to get the styling and text selection/search to work, we needed to
embed the entire SVG element on the page. I tried using `<object...>` as well,
which makes the text interactible, but it also sandboxes the styles, making it
impossible to change the light/dark mode styling of the SVG when the website
them changed.

The one piece here I really don't like is using `[data-md-color-scheme="slate"]`
in the SVG directly, since it means the SVG styling is tied to the current
website implementation. As long as we're consistent with its use, though, it
should be easy enough to search and replace if we ever need to.