Add manual PDF index
Update the options passed to pandoc and run it twice with `makeindex` in
between.

Add a Lua filter that passes through `\index` within HTML comments so that we
can custom add our index entries.

It would be nice to be able to use these index entries in the lunr.js search
results as well, but that will have to wait.